webapp/requirements.txt
boto==2.40.0
Django==1.9.7
django-rest-auth==0.7.0
django-storages==1.4.1
djangorestframework==3.3.3
hurry.filesize==0.9
image==1.5.3
Pillow==3.2.0
six==1.10.0
